% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AQSearch.CAS.R
\name{AQSearch.CAS}
\alias{AQSearch.CAS}
\title{Search CAS codes in Chemical database.}
\usage{
\method{AQSearch}{CAS}(db, ChemString, ...)
}
\arguments{
\item{db}{A list of three data.frame variables within all systems data are stored.}

\item{ChemString}{A String with a text fragment of the chemical name the user want to search the correspondent CAS.}

\item{...}{Additional optional arguments. None are used at present.}
}
\description{
Search CAS codes in Chemical database.
}
\examples{
\dontrun{
AQSearch.CAS(db, "NaOH")
}
}

