% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQSysDB.R
\name{AQSysDB}
\alias{AQSysDB}
\title{AQSysDB}
\usage{
AQSysDB(path, CAS = FALSE)
}
\arguments{
\item{path}{String containing the full path to the XLS or XLSX file.}

\item{CAS}{The user has the option to identify the component's cells in the worksheet with the CAS (CAS = TRUE) or with the row number that matches a CAS entry in the CASDB worksheet (CAS = FALSE)}
}
\description{
Import DB data from an Excel Worksheet.
}
\examples{
\dontrun{
AQSysDB("C:/data.xls", CAS = FALSE)
}
}
