% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQSysDOE.R
\name{AQSysDOE}
\alias{AQSysDOE}
\title{AQSysDOE}
\usage{
AQSysDOE(dataSET, slope = NULL, xmax = NULL, modelName = "merchuk",
  nTL = 3, nPoints = 3, tol = 1e-05)
}
\arguments{
\item{dataSET}{- Binodal Experimental data that will be used in the nonlinear fit.}

\item{slope}{The method assumes all tielines for a given ATPS are parallel, thus only one slope is required. [type:double]}

\item{xmax}{Maximum value for the Horizontal axis' value (bottom-rich component). [type:double]}

\item{modelName}{Character String specifying the nonlinear empirical equation to fit data.
The default method uses Merchuk's equation. Other mathematical descriptors can be listed using AQSysList(). [type:string]}

\item{nTL}{Number of tielines plotted for a given ATPS. Default is 3. [type:Integer]}

\item{nPoints}{Number of points chosen for a given tieline. Default is 3. [type:Integer]}

\item{tol}{limit of tolerance to reach to assume convergence. Default is 1e-5. [type:Integer]}
}
\description{
XXX.
}
