% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQSys.mathDesc.R
\name{AQSysList}
\alias{AQSysList}
\title{Aqueous Systems Descriptors already implemented}
\usage{
AQSysList(npars = FALSE)
}
\arguments{
\item{npars}{Logic option to return a List variable containing the number of
required parameters for each equation.}
}
\description{
The function returns a list of all mathematical descriptors 
available at the time.
}
