% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLSRUtils.R
\name{export_data}
\alias{export_data}
\title{LLSR Data Exporter}
\usage{
export_data(localData = NULL)
}
\arguments{
\item{localData}{A variable existing in R environment and that will be saved
locally.}
}
\description{
The function saves a copy of a specified variable to a file in 
the folder pointed by the user.
}
