% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLSRUtils.R
\name{export_template}
\alias{export_template}
\title{LLSR Template Exporter}
\usage{
export_template()
}
\description{
The function makes a copy of LLSR's template file and copy it to the folder pointed by the user.
}
