% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorporateColours.R
\name{llsr_cols}
\alias{llsr_cols}
\title{Function to extract llsr colors as hex codes}
\usage{
llsr_cols(...)
}
\arguments{
\item{...}{Character names of llsr_colors}
}
\description{
Function to extract llsr colors as hex codes
}
