% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{llsr_data}
\alias{llsr_data}
\title{DESCRIBE
DESCRIBE}
\format{DESCRIBE
\describe{
  \item{XC}{Fraction of Salt}
  \item{YC}{Fraction of PEG}
  ...
}}
\source{
\url{https://github.com/hutchr/LLSR}
}
\description{
DESCRIBE
DESCRIBE
}
\examples{
#
XYdt <- llsr_data[["db.data"]][, 93:104]
#
Xdt <- llsr_data[["db.data"]][, 93]
#
Ydt <- llsr_data[["db.data"]][, 94]
}
