% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLSRxy.R
\name{LLSRxy}
\alias{LLSRxy}
\title{Dataset aggregation Function}
\usage{
LLSRxy(FirstCol, SecondCol, ColDis = "xy")
}
\arguments{
\item{FirstCol}{- Component X's concentration in the TOP Y-rich phase.}

\item{SecondCol}{- Component Y's concentration in the BOTTOM X-rich phase.}

\item{ColDis}{Defines how the data is organized in the Worksheet. Use "xy" whether the first column corresponds to the lower phase fraction and "yx" whether the opposite.}
}
\value{
XYdt - Standard bidimensional data.frame used in most of functions available in this package.
}
\description{
This function returns a data.frame from given two unidimensionals vectors, XC and YX.
}
\details{
The function returns a data.frame after merging two unidimensional vectors.
}
\examples{
#
Xdt<-peg4kslt[,1]
#
Ydt<-peg4kslt[,1]
#
LLSRxy(Xdt,Ydt)
}
