% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mrchk.R
\name{mrchk.plot}
\alias{mrchk.plot}
\title{Dataset and Fitted Function plot}
\usage{
\method{mrchk}{plot}(XYdt, xlbl = "Salt Fraction (w/w)",
  ylbl = "PEG Fraction (w/w)", main = NULL, col = "blue", type = "p",
  cex = 2.5, cexlab = 2.5, cexaxis = 2.5, cexmain = 2.5, cexsub = 2.5,
  xmax = 0.4, ymax = 0.5, HR = FALSE, NP = 100, ...)
}
\arguments{
\item{XYdt}{- Binodal Experimental data that will be used in the nonlinear fit}

\item{xlbl}{= Plot's Horizontal axis label. If not set, It will admit the system under study is a PEG-Salt System.}

\item{ylbl}{= Plot's Vertical axis label. If not set, It will admit the system under study is a PEG-Salt System.}

\item{main}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{col}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{type}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cex}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexlab}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexaxis}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexmain}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexsub}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{xmax}{- Maximum value for the Horizontal axis' value}

\item{ymax}{- Maximum value for the Vertical axis' value}

\item{HR}{- Magnify Plot's text to be compatible with High Resolution size [type:Boulean]}

\item{NP}{- Number of points used to build the fitted curve. Default is 100. [type:Integer]}

\item{...}{Additional optional arguments. None are used at present.}
}
\value{
A plot containing the experimental data and the correspondent curve for the binodal in study.
}
\description{
The function returns a plot after fitting a dataset to the equation described by Merchuk.
}
\details{
This version uses the plot function and return a regular bidimensional plot. Future versions will include a ternary
diagram and more formal formatting.
}
\note{
\deqn{ y = P1\times{} \exp{(P2*x^1/2-P3*x^3) }}
}
\examples{
#Populating variable XYdt with binodal data
XYdt <- peg4kslt[,1:2]
#Plot XYdt using Merchuk's function
#
mrchk.plot(XYdt)
#
}

