% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xyJn.R
\name{xyJn}
\alias{xyJn}
\title{Dataset aggregation Function}
\usage{
xyJn(XC, YC, ...)
}
\arguments{
\item{XC}{- Component X's concentration in the BOTTOM X-rich phase.}

\item{YC}{- Component Y's concentration in the TOP Y-rich phase.}

\item{...}{Additional parameters - Not implemented}
}
\value{
XYdt - Standard bidimensional data.frame used in most of functions available in this package.
}
\description{
This function returns a data.frame from given two unidimensionals vectors, XC and YX.
}
\details{
The function returns a data.frame after merging two unidimensional vectors.
}
\examples{
#
Xdt<-peg4kslt[,1]
#
Ydt<-peg4kslt[,1]
#
xyJn(Xdt,Ydt)
}

