% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQSysDB.R
\name{AQSysDB}
\alias{AQSysDB}
\title{AQSysDB}
\usage{
AQSysDB(path)
}
\arguments{
\item{path}{String containing the full path to the XLS or XLSX file.}
}
\description{
Import DB data from an Excel Worksheet.
}
\examples{
\dontrun{
AQSysDB("C:/data.xls")
}
}
