% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AQSysDB.R
\name{AQSysDB}
\alias{AQSysDB}
\title{AQSysDB}
\usage{
AQSysDB(path, order = "xy", CAS = FALSE)
}
\arguments{
\item{path}{String containing the full path to the XLS or XLSX file.}

\item{order}{Defines how the data is organized in the Worksheet. Use "xy" whether the first column corresponds to the lower phase fraction and "yx" whether the opposite.}

\item{CAS}{The user has the option to identify the component's cells in the worksheet with the CAS (CAS = TRUE) or with the row number that matches a CAS entry in the CASDB worksheet (CAS = FALSE)}
}
\description{
AQSysDB
}
\examples{
\dontrun{
AQSysDB("C:/data.xls", order = "xy", CAS = FALSE)
}
}

