% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{llsr_data}
\alias{llsr_data}
\title{DESCRIBE}
\format{DESCRIBE
\describe{
  \item{XC}{Fraction of the component enriched in the lower phase}
  \item{YC}{Fraction of the component enriched in the upper phase}
  ...
}}
\source{
\url{https://github.com/hutchr/LLSR}
}
\description{
DESCRIBE
}
\examples{
#
XYdt <- llsr_data[["db.data"]][, 93:104]
#
Xdt <- llsr_data[["db.data"]][, 93]
#
Ydt <- llsr_data[["db.data"]][, 94]
}
