% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{peg4kslt}
\alias{peg4kslt}
\title{Dataset of experimental binodal data of an ATPS}
\format{multi-level data.frame()
\describe{
  \item{XC}{Ammonium Sulphate mass fraction}
  \item{YC}{Poly(ethylene glycol) mass fraction}
  ...
}}
\description{
A dataset containing the experimental binodal data for a PEG/SALT Aqueous Two-Phases System (ATPS)
}
\examples{
#
XYdt <- peg4kslt[,1:2]
#
Xdt<-peg4kslt[,1]
#
Ydt<-peg4kslt[,1]
}
