% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{peg4kslt}
\alias{peg4kslt}
\title{Dataset of experimental binodal data of an ATPS}
\format{A data.frame with 116 rows and 2 variables:
\describe{
  \item{XC}{Fraction of Salt}
  \item{YC}{Fraction of PEG}
  ...
}}
\source{
\url{http://diegofcoelho.github.io/code/}
}
\description{
A dataset containing the experimental binodal data for a PEG/SALT Aqueous Two-Phases System (ATPS)
}
\examples{
#
XYdt <- peg4kslt[,1:2]
#
Xdt<-peg4kslt[,1]
#
Ydt<-peg4kslt[,1]
}

