% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsnchk.R
\name{plotll}
\alias{plotll}
\title{This functions plot a curve based in the chosen model and its parameters.}
\usage{
plotll(model, param, xlbl = "Salt Fraction (w/w)",
  ylbl = "PEG Fraction (w/w)", main = "Title", col = "black",
  type = "p", cex = 1, cexlab = 1, cexaxis = 1, cexmain = 1,
  cexsub = 1, xmax = 0.4, mpl = FALSE, HR = FALSE)
}
\arguments{
\item{model}{- Model chosen to be used: merchuk, murugesan [type:string]}

\item{param}{- Model's parameters [type::data.frame]}

\item{xlbl}{- Plot's Horizontal axis label. If not set, It will admit the system under study is a PEG-Salt System.  [type:string]}

\item{ylbl}{- Plot's Vertical axis label. If not set, It will admit the system under study is a PEG-Salt System.  [type:string]}

\item{main}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{col}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{type}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cex}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexlab}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexaxis}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexmain}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{cexsub}{- Legacy from plot package. For more details, see \code{\link{plot.default}}}

\item{xmax}{- Maximum value for the Horizontal axis' value (bottom-rich component)  [type:double]}

\item{mpl}{- You can overlay curves in a single plot. Default is FALSE. [type::LOGIC]}

\item{HR}{- Magnify Plot's text to be compatible with High Resolution size [type:Boulean]}
}
\value{
A plot using the input model within the chosen interval. If no interval is selected, xmax=0.4.
}
\description{
The function returns a plot after using the parameters and model given by the user.
}
\details{
The function owns predefined set of equations that can be seen below and must be used, with adequated parameters,
to return a plot which represent the chosen model.
}
\examples{
\dontrun{
plotll("murugesan",as.data.frame(c(0.90,-3.48,2.92)),mpl=TRUE,col="red")
}
}

